Data.SegTree

Documentation

class Monoid f => MonoidAction f a where

appMonoid mempty = id

appMonoid (f <> g) = appMonoid f . appMonoid g

Methods

appMonoid :: f -> a -> a

data SegTree s f a

appMonoid f (x <> y) = appMonoid f x <> appMonoid f y

appMonoid (f <> g) = appMonoid f . appMonoid g

appMonoid mempty = id

Constructors

SegTree

Fields

getSegTree :: !(MVector s a)

getDualSegTree :: !(MVector s f)

sizeSegTree :: !Int

heightSegTree :: !Int

newSegTree :: (Monoid f, Unbox f, Monoid a, Unbox a, PrimMonad m) => Int -> m (SegTree (PrimState m) f a)

buildSegTree :: (Monoid f, Unbox f, Monoid a, Unbox a, PrimMonad m) => Vector a -> m (SegTree (PrimState m) f a)

O(n)

readSegTree :: (MonoidAction f a, Unbox f, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> Int -> m a

O(log n)

writeSegTree :: (MonoidAction f a, Semigroup a, Unbox a, Unbox f, PrimMonad m) => SegTree (PrimState m) f a -> Int -> a -> m ()

O(log n)

modifySegTree :: (MonoidAction f a, Semigroup a, Unbox f, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> (a -> a) -> Int -> m ()

O(log n)

mappendFromTo :: (MonoidAction f a, Monoid a, Unbox f, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> Int -> Int -> m a

mappend [l..r) O(log n)

mappendTo :: (MonoidAction f a, Monoid a, Unbox f, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> Int -> m a

mappend [0..k) O(log n)

mappendAll :: (Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> m a

mappend [0..n) O(1)

appAt :: (MonoidAction f a, Semigroup a, Unbox f, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> Int -> f -> m ()

modify f k O(log n)

appFromTo :: (MonoidAction f a, Semigroup a, Unbox f, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> Int -> Int -> f -> m ()

mapM_ (modify f) [l..r) O(log n)

evalAt :: (MonoidAction f a, Unbox f, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> Int -> f -> m ()

O(1)

pushSegTree :: (MonoidAction f a, Unbox f, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> Int -> m ()

O(1)

pullSegTree :: (Semigroup a, Unbox a, PrimMonad m) => SegTree (PrimState m) f a -> Int -> m ()

O(1)

extendToPowerOfTwo :: Int -> Int